make the bindings more modular, write readme
