+++ javascript/README.md
@@ -1,9 +1,148 @@
-# Javascript
-
-We transpile so we can use modern js. See:
-<https://cycling74.com/forums/any-plans-to-update-support-for-recent-versions-of-js#reply-58ed21d5c2991221d9ccad8c>
+# tw.gl.repl
 
 ## TODO
 
 * the functions tied to keys should be able to also alter or generate output
- somehow, so the context needs that in too.
+somehow, so the context needs that in too.
+* add some settings to the shortkeys.json for things like alphanumerical override
+
+## About
+
+This is a repl based on the excellent th.gl.texteditor. It was partly built as a
+way to learn how to take a more modern approach to building javascript code for Max,
+but it significantly extends the original functionality, to provide a more fully
+featured and configurable repl environment.
+
+## Config
+
+Basic configuration of your repl can be achieved by loading a `shortkeys.json`
+file to reconfigure it. This file is an array of objects which bind a key number
+to a function. In contrast to th.gl.editor there are no internal functions, so everything
+is defined in this file and the user can override anything.
+The config is in the following form:
+
+```json
+{
+    "bindings": [
+        {
+            "id": "execute",
+            "asciiCode": 2044,
+            "functions": [
+                "return 'run'"
+            ]
+        },
+        {
+            "id": "backspace",
+            "asciiCode": -7,
+            "functions": [
+                "ctx.backSpace()"
+            ]
+        },
+        {
+            "id": "customSpace",
+            "asciiCode": -2,
+            "functions": [
+                "myCustomFunction"
+            ]
+        }
+    ]
+}
+```
+
+As you can see there are a number of ways to define the functions that are called,
+and it is possible to call multiple functions with a single key. Functions can be
+defined as a function body in text (which will be wrapped  
+`new Function('k', 'ctx', funcString)`), it can be a function from whatever context
+is passed in (in the case of this application it is an instance of `REPLManager`),
+or it can be a reference to a custom function.
+
+## Including custom functions
+
+One of the ways to extend the repl is to attach or preload your own functions so
+you can tie them to a key.
+To make this easier the package tries to load a file called `user-repl.js`, max should
+load this up fine if it's in your path. Inside it you have access to `i.glRender`
+and `i.repl`, you also have access to a Dict of `shortkeys.json` in `sKeys`. Which
+will be stringified and passed into the repl on `init()`
+
+ Most basic usage will be something like:
+
+```javascript
+const functionOne = (k: number, ctx: {}) => {
+    return `some message`;
+  };
+i.repl.kp.preloadFunction('doSomething', functionOne);
+```
+
+You can then use this in your `shortkeys.json` app config by binding it
+to a key
+
+```json
+{
+    "bindings": [
+        {
+            "id": "pushSpace",
+            "asciiCode": -2,
+            "functions": [
+                "doSomething"
+            ]
+        }
+    ]
+}
+```
+
+Alternatively if, for some reason, you want to configure it in code rather than
+with json you could attach the function directly
+
+```javascript
+//i.repl.kp.attachFunctions(id: string, keyCode: number, funcs: Array<KeyProcessor>)
+i.repl.kp.attachFunctions("arbitraryName", -2, [functionOne])
+```
+
+which will then be run when the key is pressed. All custom function should be of
+type `KeyProcessor` and thus have the signature `function(k: number, ctx: {})`.
+
+## Alphanumeric Characters
+
+By default alphanumeric characters are treated with a special function which
+records the keypress into a text buffer for display and output. It may be the
+case that you don't want to do this because you want to attach specific functions
+to every key. You could override the default handler:
+
+```javascript
+//user-repl.js in your path
+i.repl.kp.customAlphaNum(true);
+```
+
+### Differences from th.gl.texteditor
+
+* No internal functions, everything is user defined
+* Different shortkey.json format
+* More flexible to extend
+* Slightly different max patch around javascript
+* Written in modern modular typescript code and then transpiled to max's
+ancient engine as es3 code
+* Full set of tests
+* Max bindings are all autogenerated
+
+### Developing
+
+We transpile so we can use modern js. See:
+<https://cycling74.com/forums/any-plans-to-update-support-for-recent-versions-of-js#reply-58ed21d5c2991221d9ccad8c>
+
+You will need to `npm install` inside the js folder to develop this code, as it's
+all written in typescript and needs to be transpiled
+
+## Max bindings
+
+The entrypoint into the code for max is in an autogenerated file, this makes binding
+existing functions to the max interface easier as you just need to annotate the code
+and run the generator. Functions are annotated like
+`@maxMspBinding({ draw: true, functionName: 'cursor' })` see `MaxBindings/MaxBinding.ts`
+for a full list of options. You can annotate the class as well, which is useful
+for the eg. `instanceName` field.
+
+Run `npm compile`to build the js from ts source and to generate the bindings.
+See relevant folder for templates.
+Mostly you won't need to touch this stuff, as you can extend the repl using `shortkeys.json`
+and/or `user-repl.js` for most simple use cases.
